refactor(clerk): extract FAPI serializers and centralize org claims/permissions
