+++ packages/@emulators/clerk/src/routes/fapi-environment.ts
@@ -1,0 +1,103 @@
+// Static FAPI environment config returned by GET/POST /v1/environment.
+// clerk-js reads this on Clerk.load() to learn auth methods, display, and org settings.
+export function buildEnvironmentJson(baseUrl: string): Record<string, unknown> {
+  return {
+    api_keys_settings: { user_api_keys_enabled: false, orgs_api_keys_enabled: false },
+    auth_config: {
+      object: "auth_config",
+      id: "aac_emulate",
+      single_session_mode: false,
+      claimed_at: null,
+      reverification: false,
+    },
+    commerce_settings: {
+      billing: {
+        stripe_publishable_key: null,
+        organization: { enabled: false, has_paid_plans: false },
+        user: { enabled: false, has_paid_plans: false },
+      },
+    },
+    display_config: {
+      object: "display_config",
+      id: "display_emulate",
+      after_sign_in_url: "/",
+      after_sign_out_all_url: "/",
+      after_sign_out_one_url: "/",
+      after_sign_up_url: "/",
+      after_switch_session_url: "/",
+      application_name: "Emulate",
+      branded: false,
+      captcha_public_key: null,
+      captcha_widget_type: "invisible",
+      captcha_public_key_invisible: null,
+      captcha_provider: "turnstile",
+      captcha_oauth_bypass: [],
+      home_url: baseUrl,
+      instance_environment_type: "development",
+      logo_image_url: "",
+      favicon_image_url: "",
+      preferred_sign_in_strategy: "password",
+      sign_in_url: "/sign-in",
+      sign_up_url: "/sign-up",
+      support_email: "",
+      theme: {},
+      user_profile_url: "/user",
+      clerk_js_version: "5",
+      organization_profile_url: "/organization",
+      create_organization_url: "/create-organization",
+      after_leave_organization_url: "/",
+      after_create_organization_url: "/",
+      show_devmode_warning: false,
+      terms_url: "",
+      privacy_policy_url: "",
+      waitlist_url: "",
+      after_join_waitlist_url: "",
+    },
+    maintenance_mode: false,
+    organization_settings: {
+      enabled: true,
+      max_allowed_memberships: 100,
+      force_organization_selection: false,
+      actions: { admin_delete: true },
+      domains: {
+        enabled: true,
+        enrollment_modes: ["manual_invitation", "automatic_invitation", "automatic_suggestion"],
+        default_role: "org:member",
+      },
+      slug: { disabled: false },
+      organization_creation_defaults: { enabled: true },
+    },
+    user_settings: {
+      attributes: {
+        email_address: {
+          enabled: true,
+          required: true,
+          verifications: ["email_code"],
+          used_for_first_factor: true,
+          first_factors: ["email_code"],
+          used_for_second_factor: false,
+          second_factors: [],
+          verify_at_sign_up: false,
+        },
+        password: {
+          enabled: true,
+          required: true,
+        },
+        first_name: { enabled: true, required: false },
+        last_name: { enabled: true, required: false },
+        username: { enabled: false, required: false },
+        phone_number: { enabled: false, required: false },
+        web3_wallet: { enabled: false, required: false },
+      },
+      actions: { delete_self: true, create_organization: true },
+      social: {},
+      enterprise_sso: { enabled: false, self_serve_sso: false },
+      sign_in: { second_factor: { required: false, enabled: false } },
+      sign_up: { allowlist_only: false, progressive: true, captcha_enabled: false, mode: "public", legal_consent_enabled: false },
+      password_settings: { min_length: 8, max_length: 72, disable_hibp: true, allowed_special_characters: "" },
+      passkey_settings: { allow_autofill: false, show_sign_in_button: false },
+      username_settings: { min_length: 4, max_length: 64 },
+    },
+    protect_config: { object: "protect_config", id: "protect_emulate" },
+  };
+}
