+++ packages/@emulators/clerk/src/routes/sessions.ts
@@ -7,6 +7,7 @@ import {
   parsePagination,
   sessionResponse,
   readJsonBody,
+  resolvePrimaryOrgClaims,
 } from "../route-helpers.js";
 import { getClerkStore } from "../store.js";
 import { createSessionToken } from "./oauth.js";
@@ -102,24 +103,7 @@ export function sessionRoutes({ app, store, baseUrl, tokenMap }: RouteContext): 
     const user = cs.users.findOneBy("clerk_id", session.user_id);
     if (!user) return clerkError(c, 404, "RESOURCE_NOT_FOUND", "User not found");
 
-    const memberships = cs.memberships.findBy("user_id", user.clerk_id);
-    const firstMembership = memberships[0];
-    let orgId: string | undefined;
-    let orgRole: string | undefined;
-    let orgSlug: string | undefined;
-    let orgPermissions: string[] | undefined;
-
-    if (firstMembership) {
-      const org = cs.organizations.findOneBy("clerk_id", firstMembership.org_id);
-      if (org) {
-        orgId = org.clerk_id;
-        orgRole = firstMembership.role;
-        orgSlug = org.slug;
-        orgPermissions = firstMembership.permissions;
-      }
-    }
-
-    const jwt = await createSessionToken(store, user, sessionId, baseUrl, orgId, orgRole, orgSlug, orgPermissions);
+    const jwt = await createSessionToken(store, user, sessionId, baseUrl, resolvePrimaryOrgClaims(cs, user));
 
     cs.sessions.update(session.id, { last_active_at: nowUnix() });
 
@@ -141,7 +125,7 @@ export function sessionRoutes({ app, store, baseUrl, tokenMap }: RouteContext): 
     const user = cs.users.findOneBy("clerk_id", session.user_id);
     if (!user) return clerkError(c, 404, "RESOURCE_NOT_FOUND", "User not found");
 
-    const jwt = await createSessionToken(store, user, sessionId, baseUrl);
+    const jwt = await createSessionToken(store, user, sessionId, baseUrl, resolvePrimaryOrgClaims(cs, user));
 
     cs.sessions.update(session.id, { last_active_at: nowUnix() });
